수정 - 2일차 : 형변환 제거
동일한 타입이므로 형변환 제거

+++ src/main/java/day2/Exam11.java
@@ -18,7 +18,7 @@ public class Exam11 {
 
         long num = x;
         for (int i = 0; i<n;i++){
-            answer[i] = (long)num;
+            answer[i] = num;
             num += x;
         }
         return answer;
